distributed rendering code
