+++ packages/docs/docs/lambda/cost.mdx
@@ -8,6 +8,10 @@ crumb: 'Lambda'
 
 On this page, a few strategies for optimizing the cost of using Remotion Lambda are presented.
 
+## Less memory
+
+Reducing the memory will linearly decrease the cost. Try out different values for the memory when deploying a Lambda function and see how low you can set it without experiencing a crash.
+
 ## Lower concurrency
 
 A lot of compute time is spent on warming up the Lambda function, opening browsers, uploading and downloading assets. By using less Lambda functions, less overhead is being produced which will ultimately result in lower cost, however also in slower render speeds. See the [Lambda Concurrency](/docs/lambda/concurrency) page for more information.
@@ -19,10 +23,6 @@ Not all AWS regions have the same cost. See the [AWS Lambda](https://aws.amazon.
 ## Pre-compute data
 
 Consider whether your computation will run on every invoked Lambda function. If possible, pre-compute data once and pass it as [input props](/docs/passing-props#passing-input-props-in-the-cli) to the render.
-
-## Less memory
-
-Reducing the memory will linearly decrease the cost. Try out different values for the memory when deploying a Lambda function and see how low you can set it without experiencing a crash.
 
 ## Make the render more performant
