+++ packages/docs/docs/renderer.mdx
@@ -2,10 +2,10 @@
 image: /generated/articles-docs-renderer.png
 slug: renderer
 sidebar_label: Overview
-title: "@remotion/renderer"
+title: '@remotion/renderer'
 ---
 
-import { TableOfContents } from "../components/TableOfContents/renderer";
+import {TableOfContents} from './renderer/TableOfContents';
 
 The `@remotion/renderer` package provides APIs for rendering video server-side.
 The package is also internally used by the Remotion CLI and [Remotion Lambda](/docs/lambda).
