+++ packages/docs/docs/renderer/render-media.mdx
@@ -471,6 +471,10 @@ _boolean - default `false`_
 
 <Options id="for-seamless-aac-concatenation" />
 
+### `compositionStart?`<AvailableFrom v="4.0.279"/>
+
+An option to be used only if implementing a distributed renderer, see [Distributed rendering](/docs/distributed-rendering#2-invoking-render-functions) for what to set it to.
+
 ### `onBrowserDownload?`<AvailableFrom v="4.0.137" />
 
 <Options id="on-browser-download" />
