+++ packages/docs/docs/ssr.mdx
@@ -6,7 +6,7 @@ crumb: 'The power of'
 sidebar_label: Overview
 ---
 
-import {TableOfContents} from '../components/TableOfContents/renderer';
+import {TableOfContents} from './renderer/TableOfContents';
 
 Remotion's rendering engine is built with Node.JS, which makes it easy to render a video in the cloud.
 
@@ -78,10 +78,14 @@ jobs:
           path: out/video.mp4
 ```
 
-<Step>1</Step> Commit the template to a GitHub repository.<br />
-<Step>2</Step> On GitHub, click the <code>Actions</code> tab.<br />
-<Step>3</Step> Select the <code>Render video</code> workflow on the left.<br />
-<Step>4</Step> A <code>Run workflow</code> button should appear. Click it.<br />
+<Step>1</Step> Commit the template to a GitHub repository.
+<br />
+<Step>2</Step> On GitHub, click the <code>Actions</code> tab.
+<br />
+<Step>3</Step> Select the <code>Render video</code> workflow on the left.
+<br />
+<Step>4</Step> A <code>Run workflow</code> button should appear. Click it.
+<br />
 <Step>5</Step> Fill in the props of the root component and click <code>Run workflow</code>.<br />
 <Step>6</Step> After the rendering is finished, you can download the video under <code>Artifacts</code>.<br />
